requirements.txt
# ============================================================================
# RustyBot Dependencies - RustyBot v1.38
# Python 3.10+ Required
# ============================================================================

# ----------------------------------------------------------------------------
# GUI Framework - PyQt6
# ----------------------------------------------------------------------------
# PyQt6 provides the main application framework and UI components
PyQt6==6.9.1
PyQt6-WebEngine==6.9.0

# ----------------------------------------------------------------------------
# Multimedia - Sound & Graphics
# ----------------------------------------------------------------------------
# Pygame for sound effects and audio playback
pygame==2.6.1

# ----------------------------------------------------------------------------
# Twitch Integration
# ----------------------------------------------------------------------------
# TwitchIO for bot functionality and chat integration
twitchio==3.1.0
# Fallback IRC client for Twitch chat
python-twitch-irc==1.1.0

# ----------------------------------------------------------------------------
# Configuration & Environment
# ----------------------------------------------------------------------------
# Load environment variables from .env file (API tokens, credentials)
python-dotenv==1.0.0

# ----------------------------------------------------------------------------
# HTTP & API Requests
# ----------------------------------------------------------------------------
# HTTP library for ESI API calls and EVE2Twitch integration
requests==2.32.4

# ============================================================================
# Installation Instructions:
# ============================================================================
# 
# Install all dependencies:
#   pip install -r requirements.txt
#
# Update all dependencies to latest compatible versions:
#   pip install -r requirements.txt --upgrade
#
# Create virtual environment (recommended):
#   python -m venv .venv
#   .venv\Scripts\activate  (Windows)
#   pip install -r requirements.txt
#
# ============================================================================
# Notes:
# ============================================================================
# - PyQt6-WebEngine is required for winner animation display
# - pygame is optional but recommended for sound effects
# - twitchio requires Python 3.10 or higher
# - Keep python-twitch-irc as backup for IRC fallback
# ============================================================================